Add ephemeral CI caching performance test
This test mostly targets the performance of input fingerprinting,
with all task outputs coming from the remote build cache. I've added
a bunch of libraries which I've seen used in many projects and which
add a lot of weight to the classpath to really give the fingerprinting
code a good workout. The versions of those libraries are arbitrary, the
code does not reference them. I also opted to use the small Java project
template as a base, because the difference between it and the large Java
project was not very big, but the large one takes a lot longer on the
first iteration, as it has to compile many classes and run many tests to
fill the build cache.

Signed-off-by: Stefan Oehme <st.oehme@gmail.com>