Prefer dependencies without test sources attribute
When mapping multiple entries on the classpath to a single project
dependency in Eclipse Buildship or the XML generator for .classpath
files, prefer a dependency without the "test sources" attribute, if one
exists. Otherwise, dependencies would only be visible for test sources
and not the main sources in the dependent project.

If there is more than one entry for a project dependency after applying
this rule, the actual entry is chosen in the same order as before.

Signed-off-by: Ole Osterhagen <ole@osterhagen.info>