Use CalculatedValue infrastructure to obtain ValueSource's value
When isolating the ValueSource parameters, obtaining a project(s) lock
may be required, for example to isolate a Configuration. However, it is
possible that at the same moment another thread holding the lock is
trying to obtain the same ValueSource. Prior to this commit, when a
synchronized block was used to guard the ValueSource isolation and
computation, it caused a deadlock. The CalculatedValueContainer, which
replaced the monitor, can now yield locks held by a thread if obtaining
the ValueSource block. This way, the first thread can proceed with
isolation and compute the value source safely. Then the second thread
obtains the computed value after it is available.

Perhaps a more granular locking can be employed: as the value source
computation is isolated, we can use CalculatedValueContainer to only get
the isolated parameters, and use a monitor to run the computation
itself. However, the CalculatedValueContainer itself only acquires a
ReentrantLock, it is the parameter isolation that requests the project
lock, so the extra granularity is unlikely to yield benefits.